requirements-dev.txt
# Development dependencies for testing and code quality
# Install with: pip install -r requirements-dev.txt

# Core testing framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Test utilities
coverage==7.3.2
freezegun==1.4.0  # Time mocking for testing

# Configuration
python-dotenv==1.0.0  # Environment variable loading

# Code quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Production dependencies
-r requirements.txt
